1.3 Release Commit
* Clean up/optimized GC_Manager code.
* GC monitor will now wait till player is not in combat before duing a
collection. (though you'll really only ever see this happen if you have
a low memory cap set either manually or from having a small system ram
stat that caused it to set a lower value for this)
* Added a couple new console commands:
"garbagecollector setmemcap" and "garbagecollector stats"
* setmemcap command manually sets the max memory usage cap before GC
monitor will try to do a collection during gameplay. This is set based
on how much memory the user has but can be manually set if so desired.
* stats command allows displaying some stats on memory usage and amount
of collections. Uses a custom dfLabel! :D